# Optimize CubicBezier.length() with scipy if available