A new scipy-based implementation of `CubicBezier.length()` is added and used if scipy is available. A new test checks that the result of the scipy and default implementation match.

Simple benchmark:

```
In [4]: p = svgelements.CubicBezier((3, 4), (324, 12), (32, 234), (123, 456))

In [5]: %timeit p._length_default()
1.44 s ± 5.04 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [6]: %timeit p._length_scipy()
2.5 ms ± 143 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Due to how slow the current implementation is, this test currently increase execution time by more than a 1 minute on my computer. Should I reduce the number of iteration?

Fixes #55