@tatarize 2020-11-19
I'll check it. But, it looks great.  I might prefer catching ImportError rather than ModuleNotFoundError

https://www.scivision.dev/python-importerror-vs-modulenotfound/