@abey79 2020-11-19
Oh ok, I didn't know about ImportError vs. ModuleNotFoundError -- I'll fix it tomorrow morning.

Should we make a note somewhere that this was directly copy/pasted from svgpathtools?