@tatarize 2020-11-19
It's extremely minor. You'd need to have scipy but have it broken, So it found the module but then couldn't actually load it. So it throws the other type of error.

```AssertionError: 58.67678805630189 != 58.67678799098887 within 7 places (6.531301721679483e-08 difference)```

The error for almost equal might need to be increased there. The odds are that the linear version is broken here rather than the improved version. As for the count I think I made it the Arc one that takes forever take literally 2 as the same size since it's still a long test. 

I'll likely merge this as is and then just change the import and give the test a wider margin since the current length code with lines is likely a lot more able to allow errors creep in well below margin.