@tatarize 2020-11-19
Yeah, confirmed the last little bit I needed. The errors at 1e-6 error linear averages around 3.3e-05. When linear error is put up to 1e-12 it drops to about 3.4e-9, so the error is clearly on the slower linear method. Which is all generally expected since it's largely the same as the svgpathtools stuff and any problem there would have been discovered already.